add all classes into one Stobberi package
