+++ src/main/java/Stobberi/Command.java
@@ -1,4 +1,6 @@
-import StobberiException.StobberiException;
+package Stobberi;
+
+import Stobberi.StobberiException.StobberiException;
 
 public class Command {
     private boolean isExit;
