+++ src/main/java/Stobberi/DeleteCommand.java
@@ -1,13 +1,15 @@
-public class UnmarkCommand extends Command {
+package Stobberi;
+
+public class DeleteCommand extends Command {
     private TaskList taskList;
     private int taskNumber;
-    public UnmarkCommand(TaskList taskList, int taskNumber) {
+    public DeleteCommand(TaskList taskList, int taskNumber) {
         this.taskList = taskList;
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute() {
-        taskList.unmarkTask(taskNumber);
+        taskList.delete(taskNumber);
     }
 }
