+++ src/main/java/Stobberi/ExitCommand.java
@@ -1,3 +1,5 @@
+package Stobberi;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
         super();
