+++ src/main/java/Stobberi/StobberiException/EmptyStobberiException.java
@@ -1,4 +1,4 @@
-package StobberiException;
+package Stobberi.StobberiException;
 
 public class EmptyStobberiException extends StobberiException {
     // Work in progress
