+++ src/main/java/Stobberi/StobberiException/WrongDateTimeStobberiException.java
@@ -1,4 +1,4 @@
-package StobberiException;
+package Stobberi.StobberiException;
 
 public class WrongDateTimeStobberiException extends StobberiException {
     // Work in progress
