+++ src/main/java/Stobberi/Task.java
@@ -1,3 +1,5 @@
+package Stobberi;
+
 public class Task {
     protected String description;
     protected boolean isDone;
