+++ src/main/java/Stobberi/UnmarkCommand.java
@@ -1,13 +1,15 @@
-public class DeleteCommand extends Command {
+package Stobberi;
+
+public class UnmarkCommand extends Command {
     private TaskList taskList;
     private int taskNumber;
-    public DeleteCommand(TaskList taskList, int taskNumber) {
+    public UnmarkCommand(TaskList taskList, int taskNumber) {
         this.taskList = taskList;
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute() {
-        taskList.delete(taskNumber);
+        taskList.unmarkTask(taskNumber);
     }
 }
